backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
motor==3.6.0
pymongo==4.9.2
groq==0.13.0
google-auth==2.37.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.154.0
cryptography==44.0.0
httpx==0.28.1
python-multipart==0.0.17
pdfplumber==0.11.4
python-docx==1.1.2
python-dotenv==1.0.1
